Make fields required in Add Document modal
